refresh and reorganise patches
This package relies on git-buildpackage to manage the debian patches.

Unfortunately we have to patch a plugin-path into the OpenFX-API,
which upstream maintains in a git-submodule, which in this context
translates into a subdirectory contained in the "release tarball"
but not in the main Git repository. Thus I had to split that
into a separate patch, which can be temporarily disabled whenever
refreshing the git patchqueue